게시판 및 파일 등록 완료

+++ Spring-Security/security-method/src/main/java/com/aloha/security_method/domain/Files.java
@@ -14,14 +14,14 @@ import lombok.Data;
 @Builder
 public class Files {
 
-    private int no;
+    private Long no;
     private String id;
     private String fileName;
     private String filePath;
     private Long fileSize;
     private String type;
     private String parentTable;
-    private int parentNo;
+    private Long parentNo;
     private Date createdAt;
     private Date updatedAt;
